fix: 흩어져서 찾아 보기 힘들었던 공유 타입 정리
