+++ src/styles/market/textStyles.ts
