+++ src/types/@shared/homeTypes.ts
