+++ src/types/@shared/marketTypes.ts
