--- src/types/addItemFeildsetTypes.ts
@@ -1,7 +1,0 @@
-type AddItemFeildsetNameType = 'title' | 'description' | 'price' | 'tag';
-
-export interface IAddItemFeildset {
-  subTitle: string;
-  name: AddItemFeildsetNameType;
-  placeholder: string;
-}
