--- src/types/addItemPreviewTypes.ts
@@ -1,4 +1,0 @@
-export interface IPreview {
-  id: string;
-  url: string;
-}
