--- src/types/addItemTagTypes.ts
@@ -1,4 +1,0 @@
-export interface ITag {
-  id: string;
-  content: string;
-}
